Add files via upload
These are from the NSA leak that occured duing August 2016. I recently modified them so they id a bit more of what I wanted
Changes:
added in a few things so it works in later versions of windows 10

More dlls will come up as I tweak them and change around some of the code. 
LP: Basically a connector that connects to the implant
Implant: Stick it on the computer and gives you the privileges when you connect with it via LP 

NOTE: When using make sure to add a ._ in front of .dll to hide it
*Might push out a GUI tool to make use easier